Move load_attribute to `geometry`; better logic
